bug : attendance 기능 수정
